add synchronizeClients function
